+++ src/contracts/EventFactory.sol
@@ -14,7 +14,7 @@ contract EventFactory
         m_WalletAddress = walletAddress;
     }
 
-    function createEvent(bytes32 name, bytes32 description, bytes32 image, uint date, uint8 duration, uint24 tickets, uint16 cost) public
+    function createEvent(bytes32 name, bytes32 description, bytes32 image, uint date, uint16 duration, uint24 tickets, uint16 cost) public
     {
         address newEvent = new Event(name, description, image, date, duration, tickets, cost, m_WalletAddress);
         m_Events.push(newEvent);
